function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary==2.8.6
sendgrid==6.6.0
six==1.14.0
SQLAlchemy==1.3.23
starkbank-ecdsa==1.1.0
uamqp==1.2.14
